Update: November 10, 2023 at 6:35PM
The character has been updated. The player is now able to move left and right AND up and down. The next step is to get the environment ready. Constraints to keep the user in the world will come with the next milestone. But, for now getting an environment is the goal. Due to the nature of the game this is going to be hand sketched.