Final Milestone 1 Update: November 10, 2023 at 7:47PM
The character has been updated to flip in the direction of movement. In the next milestone, I plan to make the character more animated by adding a picture of the stickman jumping when you make it jump. I have added the background, however, this is a rough sketch of what the final environment will be. The game still doesn't have restrictions, however, I plan to add them soon. Many aspects of the game are to be determined. Overall, this final Mileston 1 product is a rough sketch of the game. I'm not sure on the title and the litte aspects of the game, however, I think I'm going to be sticking to the stickman/animated idea of the game.