Update: November 15, 2023 at 7:57PM
The character has been updated to have restrictions. The character can't go past the right and left border. HOWEVER, this only applies in the first floor. I have to modify the 'border restraining' function to be manipulated and only restrain a certain way for specific floors. I also intend to add gravity, however, at the time of writing this I need to figure that part out. This version of the code is function and has restrictions for the first floor. I also added a red dot for reference of coordinates on the plane. This has helped make the restrictions. There is still much work to be done :)