Update: November 19, 2023 at 8:28PM
In this update the player has been updated to attempt jumping. I'm not sure if I'll be able to finish the gravity, power-ups, pause/resume button/screen, and the environment affecting the player, but the raw ideas have been implemented. There is a pause button in the screen, but it doesn't have a function yet in this update. The code is a bit messy, but I plan to organize the code and make it more appealing tomorrow.