Final Milestone Update: December 3, 2023 at 7:16PM
I have implemented the power-ups and tape obstacles in the game. The power-ups do their job however, the tape still needs to be tinkered to make the game end. I have also have been able to make the character go up the ladder but still have no way of making it go down so it won't float. The game is there, the specefic details need to be implemented.