Final Milestone Update: December 3, 2023 at 10:27PM
This update is the final update for this project. I have added all the functional aspects that the game needs along with cleaning the code making it more future proof to read and fix. There is a Game-Over screen and winning screen for the user to see depending on the situation that the user is in. Overall, this project has been very fun and educational.